fix: #WB-3444 folderId not setting
